Update README.md
This is an educational project. In this project, we trained a model with Convolutional Neural Network (CNN) to detect if someone wears a mask or not in this Covid-19 situation. We used about 12000+ images to train the model.

Key Word-- COVID-19, Face-Mask detection, TensorFlow, OpenCV, Keras, ImageDataGenerator, Convolutional Neural Network (CNN), Public Safety